requirements.txt
pip==19.3.1
numpy==1.17.3
tensorflow==2.0.0
Keras==2.3.1
neat-python==0.92
